chore(debugging): register with RCM
This change registers the debugger with the RCM client.

We bridge between the RCM configuration format and the debugger format
with an adapter that translates configuration events to debugger probe
events.
